Fix SlidingWindowCV bug around `gap`
The bug resulted in leakage of gap indices into
train and eval. The issue arised when `eval_mode`
was set to "uniform".
Causes:

1 gap wasn't subtracted during
shuffle in SlidingWindowCV.split()

2 relevant test test_SlidingWindowCV_shuffle was
broken due to idiotic, but subtle bug, when the
loop variable overlapped with the same variable
outside loop, which was used as ground truth